Include ThemeableMixin by default
