+++ src/mixins/Registry.ts
@@ -1,5 +1,5 @@
 import { Constructor } from '../interfaces';
-import { diffProperty, WidgetBase } from './../WidgetBase';
+import { diffProperty, Base } from './../Base';
 import WidgetRegistry from '../WidgetRegistry';
 import { reference } from './../diff';
 
@@ -11,7 +11,7 @@ export interface RegistryMixin {
 	properties: RegistryMixinProperties;
 }
 
-export function RegistryMixin<T extends Constructor<WidgetBase<any>>>(Base: T): T & Constructor<RegistryMixin> {
+export function RegistryMixin<T extends Constructor<Base<any>>>(Base: T): T & Constructor<RegistryMixin> {
 	class Registry extends Base {
 
 		@diffProperty('registry', reference)
